Reorganize project files
- Create a file (+tests) per service, e.g. album
- Keep core functionalities like client instantiation at base
